@@ -6,7 +6,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.usthe.sureness</groupId>
     <artifactId>sureness-core</artifactId>
-    <version>1.1.0</version>
+    <version>1.1.0-jdk17</version>
     <packaging>jar</packaging>
 
     <name>sureness</name>
@@ -40,13 +40,13 @@
             scm:git:https://github.com/dromara/sureness.git
         </developerConnection>
         <url>https://github.com/dromara/sureness</url>
-        <tag>1.1.0</tag>
+        <tag>1.1.0-jdk17</tag>
     </scm>
 
     <properties>
-        <java.version>1.8</java.version>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <java.version>17</java.version>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <jakarta.servlet.version>6.0.0</jakarta.servlet.version>
         <old.servlet.version>4.0.1</old.servlet.version>
@@ -57,6 +57,8 @@
         <slf4j.version>1.7.21</slf4j.version>
         <junit.version>5.7.0</junit.version>
         <easymock.version>4.0.2</easymock.version>
+        <jakarta.ws.rs.version>3.1.0</jakarta.ws.rs.version>
+        <jakarta.xml.bind-api.version>4.0.2</jakarta.xml.bind-api.version>
     </properties>
     <dependencies>
         <dependency>
@@ -69,6 +71,12 @@
             <groupId>javax.servlet</groupId>
             <artifactId>javax.servlet-api</artifactId>
             <version>${old.servlet.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>jakarta.ws.rs</groupId>
+            <artifactId>jakarta.ws.rs-api</artifactId>
+            <version>${jakarta.ws.rs.version}</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
@@ -87,6 +95,12 @@
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
             <version>${slf4j.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>jakarta.xml.bind</groupId>
+            <artifactId>jakarta.xml.bind-api</artifactId>
+            <version>${jakarta.xml.bind-api.version}</version>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>javax.xml.bind</groupId>
