@@ -6,8 +6,8 @@ import com.usthe.sureness.subject.SubjectCreate;
 import com.usthe.sureness.subject.support.SessionSubject;
 import com.usthe.sureness.util.SurenessConstant;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import java.util.List;
 
 /**
